+++ examples/base/fresh.gen.ts
@@ -8,6 +8,7 @@ import * as $api_joke from "./routes/api/joke.ts";
 import * as $greet_name_ from "./routes/greet/[name].tsx";
 import * as $index from "./routes/index.tsx";
 import * as $markdown from "./routes/markdown.tsx";
+import * as $particles from "./routes/particles.tsx";
 import * as $Counter from "./islands/Counter.tsx";
 import * as $Form from "./islands/Form.tsx";
 import type { Manifest } from "$fresh/server.ts";
@@ -20,6 +21,7 @@ const manifest = {
     "./routes/greet/[name].tsx": $greet_name_,
     "./routes/index.tsx": $index,
     "./routes/markdown.tsx": $markdown,
+    "./routes/particles.tsx": $particles,
   },
   islands: {
     "./islands/Counter.tsx": $Counter,